Implementação da indicação de mina por perto
Números e cores são utilizados para indicar quantas minas existem na vizinhança

+++ src/jcliz/github/com/modelo/Campo.java
@@ -102,8 +102,8 @@ public class Campo {
     }
 
     //conta quantas minas estão em volta do campo com uma stream
-    public long minasNaVizinanca() {
-        return vizinhos.stream().filter(v -> v.minado).count();
+    public int minasNaVizinanca() {
+        return (int) vizinhos.stream().filter(v -> v.minado).count();
     }
 
     public void reiniciar() {
